+++ apps/cli/src/main/kotlin/org/pcsoft/framework/kube/kts/cli/MainCommand.kt
@@ -36,6 +36,7 @@ import picocli.CommandLine.*
         LintCommand::class,
         TemplateCommand::class,
         InstallCommand::class,
+        UpgradeCommand::class,
         UninstallCommand::class
     ],
     header = ["Kube KTS 0.1.0 - 2026", ""],